Resolve problems around BLE mode ids
